Delta Flows and Diversions

A named list with datasets containing the inflow, total diversions, and proportion diverted for the North and South Deltas from three different Calsim runs. The components of the list contain output from four different Calsim runs:

biop_2008_2009: output from the 2008-2009 Calsim II BiOp run.

biop_itp_2018_2019: output from the 2018-2019 Calsim II BiOp run.

calsim_run_of_river: output from the Calsim Run of River run.

LTO-12a: output from the LTO-12a Calsim III run.

delta_flows

Format

a named list where each element contains a dataframe with 972 rows and 7 variables:

date

CALSIM II date

n_dlt_inflow_cfs

C400 + C157, north delta inflow in cubic feet per second

s_dlt_inflow_cfs

C401B + C504 + C508 + C644, south delta inflow in cubic feet per second

n_dlt_div_cfs

D403A + D403B + D403C + D403D + D404, north delta diversions in cubic feet per second

s_dlt_div_cfs

D418 + D419 + D412 + D410 + D413 + D409B + D416 + D408_OR + D408_VC, south delta diversions in cubic feet per second

n_dlt_prop_div

north delta diversions / north delta inflow

s_dlt_prop_div

south delta diversions / south delta inflow

Source

Data Wrangling: Sadie Gill sgill@flowwest.com

Node Selection: Mark Tompkins mtompkins@flowwest.com and Mike Urkov mike.urkov@gmail.com

CALSIM Model Output: Michael Wright mwright@usbr.gov

Details

The inflow, diversions, and proportions diverted in the North and South Deltas are represented using 'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II.

The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River below Vernalis. The nodes and calculation for each delta are outlined above.

CALSIM II schematic