Flow in cfs

A dataset containing the flow in cubic feet per second within watersheds for use with the CVPIA SIT Salmon Population Model to develop habitat inputs.

flows_cfs

Format

dataframe with 985 rows and 33 variables:

date

CALSIM II date

Upper Sacramento River

C104

Antelope Creek

C11307

Battle Creek

C10803

Bear Creek

C11001*

Big Chico Creek

C11501

Butte Creek

C217A

Clear Creek

C3

Cottonwood Creek

C10802

Cow Creek

C10801

Deer Creek

C11309

Elder Creek

C11303

Mill Creek

C11308

Paynes Creek

C11001

Stony Creek

C142A

Thomes Creek

C11304

Upper-mid Sacramento River

C115

Bear River

C285

Feather River

C203

Yuba River

C230

Lower-mid Sacramento River1

C134

Lower-mid Sacramento River2

C160

American River

C9

Lower Sacramento River

C166

Calaveras River

C92

Cosumnes River

C501

Mokelumne River

C91**

Merced River

C561

Stanislaus River

C520

Tuolumne River

C540

San Joaquin River

C630

Source

Data Wrangling: Sadie Gill sgill@flowwest.com

Node Selection: Mark Tompkins mtompkins@flowwest.com and Mike Urkov mike.urkov@gmail.com

CALSIM Model Output: Michael Wright mwright@usbr.gov

Details

The flow is represented using 'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II. The nodes and calculation for each watershed are outlined above.

The Sutter and Yolo Bypasses' flows are stored in a seperate dataset bypass_flows to enable the habitat segmentation in the bypasses according to weir locations.

The Lower-mid Sacramento River has two nodes, one above Fremont Weir (C134) and one below (C160). When calculating habitat for the Lower-Mid Sacramento river, calculate the habitat at each flow node and sum them proportion to the length of stream above and below the weir: $$35.6/58 * (habitat at C134) + 22.4/58 * (habitat at C160)$$

*Because there are no modeled flows at Bear Creek, flows at Paynes Creek are used

**Mokelumne River flow is from a separate model provided by EBMUD.

CALSIM II schematic