Proportion of Flow Diverted

A named list where each component contains a dataset with the proportion of flow diverted within watersheds for use with the CVPIA SIT Salmon Population Model. The components of the list contain output from two different Calsim II BiOp runs:

biop_2008_2009: output from the 2008-2009 Calsim II BiOp run.

biop_itp_2018_2019: output from the 2018-2019 Calsim II BiOp run.

proportion_diverted

Format

a named list of two components, each containing a 31 by 12 by 21 array watershed, month, year (1980-2000)

Source

Data Wrangling: Sadie Gill sgill@flowwest.com

Node Selection: Mark Tompkins mtompkins@flowwest.com and Mike Urkov mike.urkov@gmail.com

CALSIM Model Output: Michael Wright mwright@usbr.gov

Details

The proportion diverted was calculated using 'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II. The nodes and calculation for each watershed are outlined below

The diversions of Antelope Creek, Deer Creek and Mill Creek are represented by one diversion node. Diversions for these creeks were estimated in proportion to their flow. Elder Creek and Thomes Creek are also represented with a single node and their diversions were estimated using the same method.

Upper Sacramento River

D104 / C104

Antelope Creek

(C11307 / (C11307 + C11308 + C11309) * D11305) / C11307

Battle Creek

no modeled diversions

Bear Creek

no modeled diversions

Big Chico Creek

no modeled diversions

Butte Creek

(C217B + D217) / (C217B + D217 + C217A)

Clear Creek

no modeled diversions

Cottonwood Creek

no modeled diversions

Cow Creek

no modeled diversions

Deer Creek

(C11309 / (C11307 + C11308 + C11309) * D11305) / C11309

Elder Creek

(C11303 / (C11303 + C11304) * D11301) / C11303

Mill Creek

(C11308 / (C11307 + C11308 + C11309) * D11305) / C11308

Paynes Creek

no modeled diversions

Stony Creek

D17301 / C42

Thomes Creek

(C11304 / (C11303 + C11304) * D11301) / C11304

Upper-mid Sacramento River

(D109 + D112 + D113A + D113B + D114 + D118 + D122A + D122B + D123 + D124A + D128_WTS + D128) / C110

Sutter Bypass

no modeled diversions

Bear River

D285 / (C285 + D285)

Feather River

(D201 + D202 + D7A + D7B) / C6

Yuba River

D230 / (C230 + D230)

Lower-mid Sacramento River

(D129A + D134 + D162 + D165) / C128

Yolo Bypass

no modeled diversions

American River

D302 / C9

Lower Sacramento River

(D167 + D168 + D168A_WTS) / C166

Calaveras River

(D506A + D506B + D506C + D507) / C92

Cosumnes River

no modeled diversions

Mokelumne River

(D503A + D503B + D503C + D502A + D502B) / C91*

Merced River

(D562 + D566) / C561

Stanislaus River

D528 / C520

Tuolumne River

D545 / C540

San Joaquin River

(D637 + D630B + D630A + D620B) / (D637 + D630B + D630A + D620B + C637)

*Mokelumne River flow and diversions are from a separate model provided by EBMUD.

CALSIM II schematic