Upper Sacramento River Flow

A named list where each component contains a dataset with the Upper Sacramento River flow in cubic meters per second. The components of the list contain output from four different Calsim runs and two runs incorporating synthetic functional flow hydrographs:

biop_2008_2009: output from the 2008-2009 Calsim II BiOp run.

biop_itp_2018_2019: output from the 2018-2019 Calsim II BiOp run.

calsim_run_of_river: output from the Calsim Run of River run.

eff: synthetic functional flow hydrograph on the Upper Sacramento River and San Joaquin River.

LTO-12a: output from the LTO-12a Calsim III run

LTO_12a_eff_dy: output from the LTO-12a Calsim III run with synthetic functional flow hydrographs spliced in on the Upper Sacramento River and San Joaquin River for dry years.

upper_sacramento_flows

Format

a named list of six components, each containing a 12 by 21 matrix (month by year)

Source

Data Wrangling: Sadie Gill sgill@flowwest.com

Node Selection: Mark Tompkins mtompkins@flowwest.com and Mike Urkov mike.urkov@gmail.com

CALSIM Model Output: Michael Wright mwright@usbr.gov

Details

The Upper Sacramento River is represented using node CALSIM II 'FLOW-CHANNEL' C109 node at Bend. Each row represents a month, each column a year from 1980-2000. This data is used to route fish into the delta.

CALSIM II schematic