Proportion of Sacramento River Flow through the Sutter and Yolo Bypasses

The monthly proportion of Sacramento River flow within the bypasses (years 1980-2000) for use with the CVPIA SIT Salmon Population Model to apportion fish onto the bypasses. The components of the list contain output from four different Calsim runs:

biop_2008_2009: output from the 2008-2009 Calsim II BiOp run.

biop_itp_2018_2019: output from the 2018-2019 Calsim II BiOp run.

calsim_run_of_river: output from the Calsim Run of River run.

LTO-12a: output from the LTO-12a Calsim III run.

proportion_flow_bypasses

Format

a named list where each element contains a 12 by 21 by 2 array (month, year, bypass):

, , 1 = Sutter Bypass represented with CALSIM II nodes (D117 + D124 + D125 + D126)/C116

, , 2 = Yolo Bypass represented with CALSIM II nodes D160/C134

Source

Data Wrangling: Sadie Gill sgill@flowwest.com

Node Selection: Mark Tompkins mtompkins@flowwest.com and Mike Urkov mike.urkov@gmail.com

CALSIM Model Output: Michael Wright mwright@usbr.gov

Details

The proportions of Sacramento River flowing through the bypasses are represented using 'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II.

Some of the calculated values for Yolo Bypass were greater than 1. These values were capped at 1 for model usage.

CALSIM II schematic